+++ src/SimpleTestSuiteFactory.php
@@ -15,7 +15,7 @@ final class SimpleTestSuiteFactory implements ITestSuiteFactory
     public function create(string $className): TestCase
     {
         if (!is_subclass_of($className, TestCase::class)) {
-            throw new InvalidTestCaseException("$className is not a descendant of " . TestCase::class . ".");
+            throw new InvalidTestSuiteException("$className is not a descendant of " . TestCase::class . ".");
         }
         return new $className();
     }