+++ tests/SimpleTestSuiteFactoryTest.php
@@ -19,6 +19,6 @@ final class SimpleTestSuiteFactoryTest extends TestCase
         $this->assertType(self::class, $factory->create(self::class));
         $this->assertThrowsException(function () use ($factory) {
             $factory->create(\stdClass::class);
-        }, InvalidTestCaseException::class, "stdClass is not a descendant of " . TestCase::class . ".");
+        }, InvalidTestSuiteException::class, "stdClass is not a descendant of " . TestCase::class . ".");
     }
 }